+++ pkgs/test_api/lib/src/backend/runtime.dart
@@ -29,6 +29,12 @@ final class Runtime {
       'Safari', 'safari', Compiler.dart2js, [Compiler.dart2js],
       isBrowser: true, isJS: true);
 
+  /// Microsoft Internet Explorer.
+  @Deprecated('Internet Explorer is no longer supported')
+  static const Runtime internetExplorer = Runtime(
+      'Internet Explorer', 'ie', Compiler.dart2js, [Compiler.dart2js],
+      isBrowser: true);
+
   /// Microsoft Edge (based on Chromium).
   static const Runtime edge = Runtime(
       'Microsoft Edge', 'edge', Compiler.dart2js, [Compiler.dart2js],